Add some tests (#68)
"Look! 
 Up in the sky!" 
It's a bird!" 

"It's a plane!" 

"It's Superman!" 

🎵 It's Superman.  The Man of Steel.  Superman. 🎵 

"SUPERMAN"! 
Rocketed to Earth as an infant when the distant planet Krypton exploded! 
And who disguised as Clark Kent, 
mild mannered reporter for The Daily Planet, 
fights a never-ending battle for truth, justice and freedom, 
with super powers far beyond those of ordinary mortals. 

🎵 It's Superman.  Super, Super, Superman. 🎵



* Improve type hints for server calls

* Move settings loading to CLI

* Add context manager functions

* Add test fixtures

* Add tests for server calls

* Add tests for manager server calls

* Remove env settings from gh workflow

* Move WorkloadProducer